frontend-streamlit/pyproject.toml
[project]
name = "frontend-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "streamlit-cookies-manager>=0.2.0",
]
